pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tools.setuptools]
py_modules = ["text_me"]

[project]
name="text_me"
version="0.0.1"
dependencies = [
		"aiohttp",
		"aiohttp_jinja2",
		"aiodns",
		"pyyaml",
		"pymongo",
		"mongo",
		]

[project.scripts]
run_srv = "text_me.text_me:run" 
